Refactored Code

See merge request courses/2020-fall/csci-5308/group1/csci5308!45

+++ src/main/java/dhl/leagueModel/gamePlayConfig/GameResolver.java
@@ -3,15 +3,15 @@ package dhl.leagueModel.gamePlayConfig;
 import org.json.simple.JSONObject;
 
 public class GameResolver implements IGameResolver {
-    private int randomWinChance;
+    private double randomWinChance;
 
     @Override
-    public int getRandomWinChance() {
+    public double getRandomWinChance() {
         return randomWinChance;
     }
 
     @Override
-    public void setRandomWinChance(int randomWinChance) {
+    public void setRandomWinChance(double randomWinChance) {
         this.randomWinChance = randomWinChance;
     }
